+++ lib/active_storage_validations/shared/asv_errorable.rb
@@ -30,12 +30,19 @@ module ActiveStorageValidations
     def get_filename(file)
       return nil unless file
 
+      filename_from(file)&.to_s.presence
+    end
+
+    def filename_from(file)
       case file
-      when ActiveStorage::Attached, ActiveStorage::Attachment then file.blob&.filename&.to_s
+      when ActiveStorage::Attached, ActiveStorage::Attachment then file.blob&.filename
       when ActiveStorage::Blob then file.filename
+      when ActionDispatch::Http::UploadedFile, Rack::Test::UploadedFile then file.original_filename
       when String then ActiveStorage::Blob.find_signed!(file)&.filename
       when Hash then file[:filename]
-      end.to_s
+      when File then File.basename(file)
+      when Pathname then File.basename(file)
+      end
     end
   end
 end